+++ packages/queryable/parsers.ts
@@ -7,7 +7,6 @@ export function DefaultParse(): TimelinePipe {
     return parseBinderWithErrorCheck(async (response) => {
 
         if ((response.headers.has("Content-Length") && parseFloat(response.headers.get("Content-Length")) === 0) || response.status === 204) {
-
             return {};
         }
 
@@ -41,7 +40,7 @@ export function BufferParse(): TimelinePipe {
 export async function errorCheck(url: URL, response: Response, result: any): Promise<[URL, Response, any]> {
 
     if (!response.ok) {
-        // within these observers we just throw to indicate an unrecoverable error within the pipeline
+        // within observers we just throw to indicate an unrecoverable error within the pipeline
         throw await HttpRequestError.init(response);
     }