+++ packages/sp-addinhelpers/index.ts
@@ -1,4 +1,4 @@
-export * from "./sprequestexecutorclient.js";
+// export * from "./sprequestexecutorclient.js";
 export * from "./sprestaddin.js";