+++ packages/sp-addinhelpers/sprequestexecutorclient.ts
@@ -1,79 +1,81 @@
-import { assign, IHttpClientImpl } from "@pnp/core";
+// TODO::
 
-/**
- * Makes requests using the SP.RequestExecutor library.
- */
-export class SPRequestExecutorClient implements IHttpClientImpl {
-    /**
-     * Fetches a URL using the SP.RequestExecutor library.
-     */
-    public fetch(url: string, options: any): Promise<Response> {
-        if (SP === undefined || SP.RequestExecutor === undefined) {
-            throw Error("SP.RequestExecutor is undefined. Load the SP.RequestExecutor.js library (/_layouts/15/SP.RequestExecutor.js) before loading the PnP JS Core library.");
-        }
+// import { assign, IHttpClientImpl } from "@pnp/core";
 
-        const addinWebUrl = url.substring(0, url.indexOf("/_api")),
-            executor = new SP.RequestExecutor(addinWebUrl);
+// /**
+//  * Makes requests using the SP.RequestExecutor library.
+//  */
+// export class SPRequestExecutorClient implements IHttpClientImpl {
+//     /**
+//      * Fetches a URL using the SP.RequestExecutor library.
+//      */
+//     public fetch(url: string, options: any): Promise<Response> {
+//         if (SP === undefined || SP.RequestExecutor === undefined) {
+//             throw Error("SP.RequestExecutor is undefined. Load the SP.RequestExecutor.js library (/_layouts/15/SP.RequestExecutor.js) before loading the PnP JS Core library.");
+//         }
 
-        let headers: { [key: string]: string } = {},
-            iterator: IterableIterator<[string, string]>,
-            temp: IteratorResult<[string, string]>;
+//         const addinWebUrl = url.substring(0, url.indexOf("/_api")),
+//             executor = new SP.RequestExecutor(addinWebUrl);
 
-        if (options.headers && options.headers instanceof Headers) {
-            iterator = <IterableIterator<[string, string]>>options.headers.entries();
-            temp = iterator.next();
-            while (!temp.done) {
-                headers[temp.value[0]] = temp.value[1];
-                temp = iterator.next();
-            }
-        } else {
-            headers = <any>options.headers;
-        }
+//         let headers: { [key: string]: string } = {},
+//             iterator: IterableIterator<[string, string]>,
+//             temp: IteratorResult<[string, string]>;
 
-        return new Promise((resolve, reject) => {
+//         if (options.headers && options.headers instanceof Headers) {
+//             iterator = <IterableIterator<[string, string]>>options.headers.entries();
+//             temp = iterator.next();
+//             while (!temp.done) {
+//                 headers[temp.value[0]] = temp.value[1];
+//                 temp = iterator.next();
+//             }
+//         } else {
+//             headers = <any>options.headers;
+//         }
 
-            let requestOptions = {
-                error: (error: SP.ResponseInfo) => {
-                    reject(this.convertToResponse(error));
-                },
-                headers: headers,
-                method: options.method,
-                success: (response: SP.ResponseInfo) => {
-                    resolve(this.convertToResponse(response));
-                },
-                url: url,
-            };
+//         return new Promise((resolve, reject) => {
 
-            if (options.body) {
-                requestOptions = assign(requestOptions, { body: options.body });
-            } else {
-                requestOptions = assign(requestOptions, { binaryStringRequestBody: true });
-            }
-            executor.executeAsync(requestOptions);
-        });
-    }
+//             let requestOptions = {
+//                 error: (error: SP.ResponseInfo) => {
+//                     reject(this.convertToResponse(error));
+//                 },
+//                 headers: headers,
+//                 method: options.method,
+//                 success: (response: SP.ResponseInfo) => {
+//                     resolve(this.convertToResponse(response));
+//                 },
+//                 url: url,
+//             };
 
-    /**
-     * Converts a SharePoint REST API response to a fetch API response.
-     */
-    private convertToResponse = (spResponse: SP.ResponseInfo): Response => {
-        const responseHeaders = new Headers();
+//             if (options.body) {
+//                 requestOptions = assign(requestOptions, { body: options.body });
+//             } else {
+//                 requestOptions = assign(requestOptions, { binaryStringRequestBody: true });
+//             }
+//             executor.executeAsync(requestOptions);
+//         });
+//     }
 
-        if (spResponse.headers !== undefined) {
-            for (const h in spResponse.headers) {
-                if (spResponse.headers[h]) {
-                    responseHeaders.append(h, spResponse.headers[h]);
-                }
-            }
-        }
+//     /**
+//      * Converts a SharePoint REST API response to a fetch API response.
+//      */
+//     private convertToResponse = (spResponse: SP.ResponseInfo): Response => {
+//         const responseHeaders = new Headers();
 
-        // Cannot have an empty string body when creating a Response with status 204
-        const body = spResponse.statusCode === 204 ? null : spResponse.body;
+//         if (spResponse.headers !== undefined) {
+//             for (const h in spResponse.headers) {
+//                 if (spResponse.headers[h]) {
+//                     responseHeaders.append(h, spResponse.headers[h]);
+//                 }
+//             }
+//         }
 
-        return new Response(body, {
-            headers: responseHeaders,
-            status: spResponse.statusCode,
-            statusText: spResponse.statusText,
-        });
-    };
-}
+//         // Cannot have an empty string body when creating a Response with status 204
+//         const body = spResponse.statusCode === 204 ? null : spResponse.body;
+
+//         return new Response(body, {
+//             headers: responseHeaders,
+//             status: spResponse.statusCode,
+//             statusText: spResponse.statusText,
+//         });
+//     };
+// }